% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{get_package_root}
\alias{get_package_root}
\title{get_package_root}
\usage{
get_package_root()
}
\value{
string The directory of this package
}
\description{
Returns the install location of the currently loaded package.
}
\author{
Alexander Vowinkel
}

