% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/environment.R
\name{get_www_dir}
\alias{get_www_dir}
\title{get_www_dir}
\usage{
get_www_dir(file = NA)
}
\arguments{
\item{file}{optional relative file in www dir to append}
}
\value{
string Location of the www ressources for shiny
}
\description{
Returns the location of the www ressources for shiny.
}
\author{
Alexander Vowinkel
}

