% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io_inner.R
\name{io_inner}
\alias{io_inner}
\title{io_inner}
\usage{
io_inner(input, output, session)
}
\value{
The inner server function for this application.
}
\description{
The inner server function, loaded inside a shiny server
}
\author{
Alexander Vowinkel
}
\seealso{
io_outer
}

