% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/io_outer.R
\name{io_outer}
\alias{io_outer}
\title{io_outer}
\usage{
io_outer()
}
\value{
The outer server function for this application.
}
\description{
The outer server function, loaded before starting a shiny server.
}
\author{
Alexander Vowinkel
}
\seealso{
io_iner
}

