% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/menu.rename.R
\name{menu.rename}
\alias{menu.rename}
\title{menu.rename}
\usage{
menu.rename(dir_initial, fbp, befCIP, ppl, yeardate, autor,
  updateProgress = NULL)
}
\arguments{
\item{dir_initial}{}

\item{fbp}{}

\item{befCIP}{}

\item{ppl}{}

\item{yeardate}{}

\item{autor}{}
}
\value{
character a string; default is 'Hello, world!'
}
\description{
The typical hello world example.
}
\author{
Mirella Flores

Alexander Vowinkel
}

