% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ui_get_page.R
\name{ui_get_page}
\alias{ui_get_page}
\title{ui_get_page}
\usage{
ui_get_page()
}
\value{
A UI defintion that can be passed to the shinyUI function.
}
\description{
To load a shiny page into a shiny application.
}
\author{
Alexander Vowinkel
}

